# internal/runtime/atomic: add arm native implementations of And8/Or8